requirements.txt
streamlit
langchain
langchain-openai
python-dotenv
faiss-cpu
pypdf
beautifulsoup4